refactor: introduced new CmdGenerators
The `DefaultCmdGenerator` was split into a `BaseCmdGenerator` and pair
of implementations for running jars and jshell scripts.
