+++ src/main/java/dev/jbang/source/CmdGenerator.java
@@ -3,5 +3,5 @@ package dev.jbang.source;
 import java.io.IOException;
 
 public interface CmdGenerator {
-	String generate(Code code, RunContext ctx) throws IOException;
+	String generate() throws IOException;
 }
